ops:forge fmt

+++ src/ChallengeContract.sol
@@ -1,40 +1,39 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity 0.8.15;
 
-import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
-import {Initializable} from "@openzeppelin/contracts/proxy/utils/Initializable.sol";
-import {ISemver} from "src/universal/ISemver.sol";
-import {NodeManager} from "src/NodeManager.sol";
-import {CommitmentManager} from "src/CommitmentManager.sol";
-import {Hashing} from "src/libraries/Hashing.sol";
-import {Verifier} from "src/kzg/Verifier.sol";
-import {Pairing} from "src/kzg/Pairing.sol";
+import { Ownable } from "@openzeppelin/contracts/access/Ownable.sol";
+import { Initializable } from "@openzeppelin/contracts/proxy/utils/Initializable.sol";
+import { ISemver } from "src/universal/ISemver.sol";
+import { NodeManager } from "src/NodeManager.sol";
+import { CommitmentManager } from "src/CommitmentManager.sol";
+import { Hashing } from "src/libraries/Hashing.sol";
+import { Verifier } from "src/kzg/Verifier.sol";
+import { Pairing } from "src/kzg/Pairing.sol";
 
-    struct Challenge {
-        uint nonce;
-        uint8 status;
-        address challenger;
-        address storageAddr;
-        address user;
-        uint start;
-        uint end;
-        uint r;
-        Pairing.G1Point aggregateCommitment;
-        uint point;
-        uint timeoutBlock;
-    }
+struct Challenge {
+    uint256 nonce;
+    uint8 status;
+    address challenger;
+    address storageAddr;
+    address user;
+    uint256 start;
+    uint256 end;
+    uint256 r;
+    Pairing.G1Point aggregateCommitment;
+    uint256 point;
+    uint256 timeoutBlock;
+}
 
-    struct ChallengeDetatils {
-        uint consensusIndex;
-        uint noConsensusIndex;
-        uint currentIndex;
-        Pairing.G1Point consensusCommitment;
-        Pairing.G1Point noConsensusCommitment;
-        Pairing.G1Point currAggregateCommitment;
-    }
+struct ChallengeDetatils {
+    uint256 consensusIndex;
+    uint256 noConsensusIndex;
+    uint256 currentIndex;
+    Pairing.G1Point consensusCommitment;
+    Pairing.G1Point noConsensusCommitment;
+    Pairing.G1Point currAggregateCommitment;
+}
 
 contract ChallengeContract is Initializable, ISemver, Ownable {
-
     /// @notice Semantic version.
     /// @custom:semver 0.1.0
     string public constant version = "0.1.0";
@@ -45,18 +44,20 @@ contract ChallengeContract is Initializable, ISemver, Ownable {
 
     Verifier public verifier;
 
-    mapping(uint => Challenge) public challenges;
+    mapping(uint256 => Challenge) public challenges;
 
-    mapping(uint => ChallengeDetatils) public challengeDetailsMap;
+    mapping(uint256 => ChallengeDetatils) public challengeDetailsMap;
 
-    uint public nonce;
+    uint256 public nonce;
 
-    event ChallengeCreated(uint nonce, address storageAddr, address user, uint start, uint end, uint r, uint timeoutBlock);
+    event ChallengeCreated(
+        uint256 nonce, address storageAddr, address user, uint256 start, uint256 end, uint256 r, uint256 timeoutBlock
+    );
 
-    event AggregateCommitmentUploaded(uint nonce, Pairing.G1Point aggregateCommitment, uint timeoutBlock);
+    event AggregateCommitmentUploaded(uint256 nonce, Pairing.G1Point aggregateCommitment, uint256 timeoutBlock);
 
     /// @notice Constructs the ChallengeContract contract.
-    constructor() {}
+    constructor() { }
 
     /// @notice Initializer
     function initialize(NodeManager _nodeManager, CommitmentManager _commitmentManager) public initializer {
@@ -65,8 +66,17 @@ contract ChallengeContract is Initializable, ISemver, Ownable {
         _transferOwnership(tx.origin);
     }
 
-
-    function createChallenge(uint _start, uint _end, address _storageAddr, address _user, uint _r, uint _point) public payable {
+    function createChallenge(
+        uint256 _start,
+        uint256 _end,
+        address _storageAddr,
+        address _user,
+        uint256 _r,
+        uint256 _point
+    )
+        public
+        payable
+    {
         require(msg.value > 0, "ChallengeContract: minimum entry fee required");
 
         Challenge storage challenge = challenges[nonce];
@@ -82,14 +92,12 @@ contract ChallengeContract is Initializable, ISemver, Ownable {
         challenge.point = _point;
         challenge.timeoutBlock = block.number + 600;
 
-
         emit ChallengeCreated(nonce, _storageAddr, _user, _start, _end, _r, challenge.timeoutBlock);
 
-        nonce ++;
-
+        nonce++;
     }
 
-    function uploadAggregateCommitment(uint _challengeId, Pairing.G1Point calldata _commitment) public {
+    function uploadAggregateCommitment(uint256 _challengeId, Pairing.G1Point calldata _commitment) public {
         Challenge storage challenge = challenges[_challengeId];
 
         require(challenge.status == 0 || challenge.status == 3, "ChallengeContract: challenge is already complete");
@@ -101,7 +109,6 @@ contract ChallengeContract is Initializable, ISemver, Ownable {
         if (challenge.status == 0) {
             challenge.aggregateCommitment = _commitment;
             challenge.status = 1;
-
         } else {
             ChallengeDetatils storage details = challengeDetailsMap[challenge.nonce];
             details.currAggregateCommitment = _commitment;
@@ -111,10 +118,9 @@ contract ChallengeContract is Initializable, ISemver, Ownable {
         emit AggregateCommitmentUploaded(_challengeId, _commitment, challenge.timeoutBlock);
 
         challenge.timeoutBlock = block.number + 600;
-
     }
 
-    function uploadProof(uint _challengeId, Pairing.G1Point calldata _proof, uint256 _value) public {
+    function uploadProof(uint256 _challengeId, Pairing.G1Point calldata _proof, uint256 _value) public {
         Challenge storage challenge = challenges[_challengeId];
 
         require(challenge.status == 5, "ChallengeContract: consensus on commitment not reached");
@@ -126,10 +132,9 @@ contract ChallengeContract is Initializable, ISemver, Ownable {
         bool isValid = verifyAggregateCommitment(challenge.aggregateCommitment, _proof, challenge.point, _value);
 
         challenge.status = isValid ? 8 : 7;
-
     }
 
-    function submitOpinion(uint _challengeId, bool _agreed) public {
+    function submitOpinion(uint256 _challengeId, bool _agreed) public {
         Challenge storage challenge = challenges[_challengeId];
 
         require(challenge.challenger == msg.sender, "ChallengeContract: only the challenger can submit an opinion");
@@ -144,11 +149,9 @@ contract ChallengeContract is Initializable, ISemver, Ownable {
             handleOngoingStatus(challenge, _agreed);
         }
         challenge.timeoutBlock = block.number + 600;
-
     }
 
     function handleInitialStatus(Challenge storage challenge, bool _agreed) internal {
-
         if (_agreed) {
             challenge.status = 5;
         } else {
@@ -178,40 +181,68 @@ contract ChallengeContract is Initializable, ISemver, Ownable {
             if (details.consensusIndex == challenge.start) {
                 details.consensusCommitment = aggregateCommitment(challenge.start, challenge.user, challenge.r);
             }
-            bool isValid = verifyAggregateCommitment(details.consensusCommitment, details.noConsensusCommitment, details.noConsensusIndex, challenge.user, challenge.r);
+            bool isValid = verifyAggregateCommitment(
+                details.consensusCommitment,
+                details.noConsensusCommitment,
+                details.noConsensusIndex,
+                challenge.user,
+                challenge.r
+            );
             challenge.status = isValid ? 8 : 7;
         }
 
         details.currentIndex = (details.noConsensusIndex + details.consensusIndex) / 2;
     }
 
-    function verifyAggregateCommitment(Pairing.G1Point memory _consensusCommitment, Pairing.G1Point memory _aggregateCommitment, uint _index, address _user, uint256 _r) public view returns (bool){
-
+    function verifyAggregateCommitment(
+        Pairing.G1Point memory _consensusCommitment,
+        Pairing.G1Point memory _aggregateCommitment,
+        uint256 _index,
+        address _user,
+        uint256 _r
+    )
+        public
+        view
+        returns (bool)
+    {
         bytes32 hash = Hashing.hashFold(_r, _index);
         Pairing.G1Point memory n0 = Pairing.mulScalar(commitment(_index, _user), uint256(hash));
         Pairing.G1Point memory n1 = Pairing.plus(_consensusCommitment, n0);
         return Pairing.equal(_aggregateCommitment, n1);
     }
 
-    function verifyAggregateCommitment(Pairing.G1Point memory _commitment,
+    function verifyAggregateCommitment(
+        Pairing.G1Point memory _commitment,
         Pairing.G1Point memory _proof,
         uint256 _index,
-        uint256 _value) public view returns (bool) {
+        uint256 _value
+    )
+        public
+        view
+        returns (bool)
+    {
         return verifier.verify(_commitment, _proof, _index, _value);
     }
 
-    function aggregateCommitment(uint _index, address _user, uint _r) public view returns (Pairing.G1Point memory){
+    function aggregateCommitment(
+        uint256 _index,
+        address _user,
+        uint256 _r
+    )
+        public
+        view
+        returns (Pairing.G1Point memory)
+    {
         Pairing.G1Point memory n0 = commitment(_index, _user);
         bytes32 hash = Hashing.hashFold(_r, _index);
         return Pairing.mulScalar(n0, uint256(hash));
     }
 
-    function commitment(uint _index, address _user) public view returns (Pairing.G1Point memory){
+    function commitment(uint256 _index, address _user) public view returns (Pairing.G1Point memory) {
         return commitmentManager.getUserCommitments(_user, _index);
     }
 
     function SetKZG(address _addr) external onlyOwner {
         verifier = Verifier(_addr);
     }
-
 }
